refactor(project): removed common entityIds
