--- commons/src/main/kotlin/piperkt/common/id/UserId.kt
@@ -1,3 +1,0 @@
-package piperkt.common.id
-
-class UserId(val value: String)
