+++ friendships-service/src/main/kotlin/piperkt/services/friendships/application/FriendshipRepository.kt
@@ -1,8 +1,8 @@
 package piperkt.services.friendships.application
 
 import piperkt.common.ddd.Repository
-import piperkt.common.id.FriendshipId
 import piperkt.services.friendships.domain.Friendship
+import piperkt.services.friendships.domain.FriendshipId
 
 interface FriendshipRepository : Repository<FriendshipId, Friendship> {
     fun findByMembers(from: String, to: String): Friendship?
