+++ friendships-service/src/main/kotlin/piperkt/services/friendships/domain/Message.kt
@@ -2,7 +2,6 @@ package piperkt.services.friendships.domain
 
 import java.time.Instant
 import piperkt.common.ddd.Entity
-import piperkt.common.id.MessageId
 
 class Message(
     override val id: MessageId = MessageId(),
