+++ friendships-service/src/main/kotlin/piperkt/services/friendships/domain/MessageId.kt
@@ -1,0 +1,5 @@
+package piperkt.services.friendships.domain
+
+import piperkt.common.ddd.UUIDEntityId
+
+class MessageId(value: String = newId()) : UUIDEntityId(value)
