+++ servers-service/src/main/kotlin/piperkt/services/servers/interfaces/web/api/dto/MessageDTO.kt
@@ -1,8 +1,8 @@
 package piperkt.services.servers.interfaces.web.api.dto
 
 import io.micronaut.serde.annotation.Serdeable
-import piperkt.common.id.MessageId
 import piperkt.services.servers.domain.Message
+import piperkt.services.servers.domain.MessageId
 
 @Serdeable
 data class MessageDTO(
